+++ src/dns/mod.rs
@@ -8,6 +8,7 @@ pub mod dnssec;
 pub mod doh;
 pub mod handler;
 pub mod homograph;
+pub mod mdns;
 pub mod resolve;
 pub mod secondary;
 pub mod server;